moviementando produto e variacoes

+++ src/main/java/com/easyerp/model/dto/ProdutoVariacaoResponse.java
@@ -7,7 +7,7 @@ import java.util.stream.Collectors;
 import com.easyerp.domain.entidade.ProdutoVariacao;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
-public record ProdutoVariacaoResponse(Long id, String descricao, String ean13, String unidade ,BigDecimal qtdeEstoque, BigDecimal qtdeporEmbalagem,  String caracterisca,
+public record ProdutoVariacaoResponse(Long id, String descricao, String ean13, String unidade ,Integer qtdeEstoque, BigDecimal qtdeporEmbalagem,  String caracterisca,
 		@JsonIgnore
 	Set<AtributoResponse> caracteristicas , Set<ComponenteResponse>componentes) {
 	public ProdutoVariacaoResponse(ProdutoVariacao produtoVariacao) {
